+++ public/documentation/js/State.js
@@ -80,6 +80,12 @@ class State {
     }
 
     setSelection(selection, hash) {
+        if (!selection) {
+            return this
+        }
+        if (selection.includes('#')) {
+            [selection, hash] = selection.split('#')
+        }
         //console.log(this.clazz, 'setSelection', selection, hash)
         selection = (this.data.map && this.data.map.hasOwnProperty(selection)) ? selection : NavTools.getRoot()
         if (selection === this.selection && !this.refresh)